Fix antora/hack/local build scripts
They needed an extra ../ inserted because of the recent directory
change.

Also add some more Makefile targets to run them and other build
related tasks.

As per commentary, I haven't figured out how to make the hugo build
output work correctly when loaded from file:// urls. So it's better
to use `hugo server` for looking at it. I'm assuming it can be fixed
in future, so I kept the build command. Perhaps the solution is to
use a simple nginx container to preview, or to wire up the hugo
server so it can see the antora docs.