Use shared file with symlinks for GitHub icon svg
The symlink technique seems like it has potential, but it's a little tricky
since there are two different templating languages being used. In
this case the shared file doesn't have any templating in it, so it
works fine.

(Could probably be squashed with previous commit, but I decided to
keep it separate to better show how the symlink was done.)