Remove taller navbar home page customizations
Adjusting this height was awkward and I don't think it was very
popular anyhow. Let's take it out and use the standard navbar height
from the antora default theme.

(Done with advanced knowledge of some incoming homepage
customizations from #46 that are going to look better with the
regular navbar height.)