ADD generation of antora header nav from Hugo TOML
This commit satisfies HACBS-2110 and allows the navigation menu for the
website to be defined in the `hugo.toml` file.
Note: Any links that are not part of the Hugo website, such as links to
any portion of the Antora generated site, need to be under the key `url`
instead of `pageRef` as `pageRef` is solely for Hugo generated pages.

SVG graphics can be used as menu content if the svg content is entered
in the `graphic` key within the menu entry in the `hugo.toml` file.

Signed-off-by: Rob Nester <rnester@redhat.com>