Add Antora built documentation as static resource
This way when running `hugo serve` the Antora built documentation is
available on `/docs` as well.
Since the output of the Antora built is nested within the
`website/antora/docs` and then copied over by Hugo to `public/docs` the
two build steps and artifacts would no longer work on GitHub Workflows,
so this also unifies them to a single build step.