Refactor to reduce number of static file symlinks
Since everything in website/static was a symlink to a file in
common, we can make things much simpler.

Originally I was using a symlink at website/static. That actually
that doesn't work in hugo, but adding to staticDir in hugo.toml
achieves the same result in a better way.