Add download button to the front page
By default the download button links to to the rolling release[1] on
GitHub, with JavaScript enabled and given support for the
OS/architecture the link is changed to the appropriate release asset and
an "All downloads" link is added to point to the rolling releases page
on GitHub.

Partial for the footer scripts is now generic and allows adding
additional scripts to the `website/assets/js` directory.

Ref. https://issues.redhat.com/browse/HACBS-2214

[1] https://github.com/enterprise-contract/ec-cli/releases/tag/snapshot