Tweak max height for zoomed in image
Without a max height, depending on their aspect ration, some images
can be taller than the screen when zoomed in, which is not a good
UX. Also you can't see the caption. Keeping some room for the
caption is why I set it to 92vh instead of 100vh.

Also made the backdrop slightly darker.