FIX broken 404s
This commit addresses RHTAPBUGS-344 and resolves issues with broken 404s
on the enterprisecontract.dev website. This commit implements a config
directory approach for hugo which allows for differing values of the
`relativeURLs` configuration key for PR preview and production
deployment. Updates were made to the Makefile to reflect the shift from
a single `hugo.toml` file to the config directory. Additionally, updates
were made to the menu parsing helper used in the Antora portion of the
website based on the implementation of the config directory approach.

Signed-off-by: Rob Nester <rnester@redhat.com>