Fixes for blog post pagination
The default pagination was set to 2, presumably because that's what
we want for the home page blog post excerpts, but the main blog page
was also using that value, which I don't think was intended.

Also I removed the pagination elements from the home page exerpts
since they were a little broken and probably not useful anyhow. I
think this should also close KONFLUX-2101 incidentally.

Ref: https://issues.redhat.com/browse/EC-491
Ref: https://issues.redhat.com/browse/KONFLUX-2101

Co-authored-by: Zoran Regvart <zoran@regvart.com>