Avoid ping-pong local rebuild triggering
I'm running `make hugo-server` in one terminal and `make
antora-local-live` in another, and I'm seeing antora rebuilds
trigger a hugo refresh which then triggers another antora rebuild and
so-on forever.

Fix it by narrowing the list of directories being watched by entr in
the antora local list script. Watch just "./antora" dir in this
repo, instead of "./" so files touched by hugo don't trigger an
antora rebuild.

(This is a quick hack - I expect there's a nicer way to do this.)

Ref: https://issues.redhat.com/browse/EC-155