Include all doc components in left panel nav
The default antora UI theme includes nav links from just the current
component. Change it so we include nav links from all the
components.

This improves the coherence of the docs and helps to reveal the
shortcomings of the current structure, which I want to try to
improve in some future PRs.

Also remove some links from the nav.adoc in this repo, since a.
they're no longer useful now that the same links are in the nav
menu, and b. because they are duplicates of the main component link,
they actually break the javascript interactive menu tree handling.

Also includes some custom CSS to tweak the size, spacing, and add some
color to highlighted nav links.

Ref: https://issues.redhat.com/browse/EC-155