Support ordering the components in the nav menu
Requires some javascript since it appears the order of the
components in `site.components` is not based on how they're listed
in the playbook file.

I'm not really sure how the order was determined previously, but
after this patch we can definitely control it.

Note that #each works roughly the same for an object and a list so
that's why there's not much changing in the partial.

Ref: https://issues.redhat.com/browse/EC-903