logo: Fix small gaps in svg source
Tweak the shape of the half background fill to remove little gaps
between it and the stroke edge of the main shape. Also move it a
couple of pixels left so it's more centered.

I did it manually for all three shapes, so they're likely not all
entirely identical now. It would be better if they were identical,
but given my lack of Inkscape experience, I think this is fine.