logo: Add initial "conforma" logo to source svg
Not sure if this is a final draft of placeholder. I learned some new
tricks in Inkscape while working on this, but (disclaimer) IANA
graphic artist.

In Inkscape:
- Picked a font that I thought looked okay and created the text
- Split up the text and did some hand kerning on the text
- Added a white background instead of transparent so the same logo
  can be used on a light or dark background.
- Exported some png files
In Gimp:
- Resized the canvas on the exported pngs to make a square

To do the kerning I needed to tweak the inner circles of the "o" and
"a" chars to avoid those letters appearing solid. At first I tried
moving the layers up and down, which worked, but didn't provide the
right transparency values when exporting. I ending up using path
difference to combine them into a single path for the letter.

Ref: https://issues.redhat.com/browse/EC-1073