Use a var for cloudflare preview project name
I'm setting up a new CloudFlare account and a new token and trying
to get the pre-merge preview working with wrangler.

I set the var at the org level to "conforma" in the new org and
"enterprise-contract" in the existing org.

What I did to set up the Cloudflare account and token:
* Sign up via https://pages.cloudflare.com/
* Create a token. Use the permissions preset as suggested
  at https://github.com/cloudflare/wrangler-action/issues/233
* Create a GitHub org secret for it called CLOUDFLARE_API_TOKEN
  at the org level, (but limited it to just this repo).
* Create a "pages" project in Cloudflare by clicking "+ Add", then
  "Pages". Name it "conforma", then set a org var called
  CLOUDFLARE_PROJECT_NAME, also limited to just this repo.