Rename to conforma in package.json
I don't think this matters much, but let's be consistent. IIUC the
package.json is just to specify the dependencies for wrangler which
does the pre-merge previews on Cloudflare.

Note: I realise it would be better to use an upstream bugs url
instead of the Jira like, but I don't want to deal with changing
that now, and also I don't think it's very important.